+++ docs/Changelog.md
@@ -12,7 +12,7 @@
  * MessageEncodingException on message compression failure (@sirn-se)
  * Remove deprecated code (@sirn-se)
  * Nyholm PSR HTTP (deprecate internal stub classes) (@sirn-se)
-
+ * getConnections() return Connections collection instead of array (@sirn-se)
 
 ## `v3.8`